refactor: split prepareT render pipeline
