+++ packages/next/src/server-dir/buildtime/__tests__/T.test.tsx
@@ -1,20 +1,26 @@
 import { describe, expect, it, vi } from 'vitest';
 
-const { mockGetRequestConditions, mockRscT } = vi.hoisted(() => ({
-  mockGetRequestConditions: vi.fn(),
-  mockRscT: vi.fn(),
-}));
+const { mockGetRequestConditions, mockRscT, mockRenderPreparedT } = vi.hoisted(
+  () => ({
+    mockGetRequestConditions: vi.fn(),
+    mockRscT: vi.fn(),
+    mockRenderPreparedT: vi.fn(),
+  })
+);
 
 vi.mock('../../../request/getRequestConditions', () => ({
   getRequestConditions: mockGetRequestConditions,
 }));
 
 vi.mock('gt-react/context', () => ({
+  createRenderPipeline: vi.fn(() => ({
+    renderPreparedT: mockRenderPreparedT,
+  })),
   T: mockRscT,
 }));
 
 describe('buildtime T', () => {
-  it('passes request conditions to RscT', async () => {
+  it('passes request conditions to context T', async () => {
     mockGetRequestConditions.mockResolvedValue({
       _locale: 'fr',
       _enableI18n: false,
@@ -32,6 +38,7 @@ describe('buildtime T', () => {
       id: 'greeting',
       _locale: 'fr',
       _enableI18n: false,
+      _renderPreparedT: expect.any(Function),
     });
     expect(T._gtt).toBe('translate-server');
   });
