+++ packages/react-core/src/components-rsc.ts
@@ -5,12 +5,11 @@
 
 export { Branch, GtInternalBranch } from './components/branches/Branch';
 export { Derive, GtInternalDerive } from './components/derivation/Derive';
+export { GtInternalVar, Var } from './components/variables/Var';
 export {
   RscGtInternalPlural as GtInternalPlural,
   RscPlural as Plural,
 } from './components/branches/Plural.rsc';
-export { RscT, RscT as T } from './components/translation/T';
-export { GtInternalVar, Var } from './components/variables/Var';
 export {
   RscCurrency as Currency,
   RscGtInternalCurrency as GtInternalCurrency,
@@ -27,5 +26,58 @@ export {
   RscGtInternalRelativeTime as GtInternalRelativeTime,
   RscRelativeTime as RelativeTime,
 } from './components/variables/RelativeTime.rsc';
-export { default as getPluralBranch } from './utils/plurals/getPluralBranch';
-export type { RelativeTimeFormatOptions, RenderVariable } from './utils/types';
+export { RscT as T } from './components/translation/T.rsc';
+
+// ===== Functions ===== //
+export { getTranslationsSnapshot } from './functions/helpers/getTranslationsSnapshot';
+export { t } from './functions/translation/t';
+
+// ===== Helpers ===== //
+export { getFormatLocales, getPluralBranch } from './pure';
+export { getShouldTranslate } from './hooks/utils/getShouldTranslate';
+export { prepareT } from './utils/translation/prepareT.shared';
+export { createRenderVariable } from './utils/rendering/createRenderVariable';
+export { createRenderPipeline } from './utils/rendering/createRenderPipeline';
+// Pre-instantiated RSC render pipeline: bound to the RSC renderVariable, so
+// consumers never thread a variable renderer through rendering calls.
+export {
+  renderDefaultChildren,
+  renderPreparedT,
+  renderTranslatedChildren,
+  renderVariable,
+} from './utils/rendering/renderPipeline.rsc';
+
+// ===== Internal ===== //
+export { internalInitializeGTSRA } from './setup/initializeGTSRA';
+export { getReadonlyConditionStoreWithFallback } from './condition-store/singleton-operations';
+export {
+  getReactI18nCache,
+  setReactI18nCache,
+} from './i18n-cache/singleton-operations';
+
+// ===== Types ===== //
+export type {
+  JsxTranslationOptions,
+  PreparedT,
+} from './utils/translation/prepareT.shared';
+export type {
+  PluralProps,
+  ResolvedPluralProps,
+} from './components/branches/Plural.shared';
+export type {
+  CurrencyProps,
+  ResolvedCurrencyProps,
+} from './components/variables/Currency.shared';
+export type {
+  DateTimeProps,
+  ResolvedDateTimeProps,
+} from './components/variables/DateTime.shared';
+export type {
+  NumProps,
+  ResolvedNumProps,
+} from './components/variables/Num.shared';
+export type {
+  RelativeTimeProps,
+  ResolvedRelativeTimeProps,
+} from './components/variables/RelativeTime.shared';
+export type { RelativeTimeFormatOptions, RenderVariable } from './pure';
