+++ packages/react-core/src/components/branches/Plural.rsc.tsx
@@ -3,7 +3,7 @@ import { renderPlural, type ResolvedPluralProps } from './Plural.shared';
 
 // RSC implementation: request conditions are passed explicitly instead of
 // being read from hooks. This module must stay free of hook/context imports
-// so it can be exported from the context-rsc entrypoint.
+// so it can be exported from the components-rsc entrypoint.
 
 function RscGtInternalPlural(props: ResolvedPluralProps): ReactNode {
   return renderPlural(props);
