+++ packages/react-core/src/components/variables/Currency.rsc.tsx
@@ -2,7 +2,7 @@ import { computeCurrency, type ResolvedCurrencyProps } from './Currency.shared';
 
 // RSC implementation: request conditions are passed explicitly instead of
 // being read from hooks. This module must stay free of hook/context imports
-// so it can be exported from the context-rsc entrypoint.
+// so it can be exported from the components-rsc entrypoint.
 
 function RscGtInternalCurrency(props: ResolvedCurrencyProps): string | null {
   return computeCurrency(props);
