+++ packages/react-core/src/components/variables/RelativeTime.rsc.tsx
@@ -5,7 +5,7 @@ import {
 
 // RSC implementation: request conditions are passed explicitly instead of
 // being read from hooks. This module must stay free of hook/context imports
-// so it can be exported from the context-rsc entrypoint.
+// so it can be exported from the components-rsc entrypoint.
 
 function RscGtInternalRelativeTime(
   props: ResolvedRelativeTimeProps
