+++ packages/react-core/src/context.ts
@@ -2,7 +2,8 @@
 export { Branch } from './components/branches/Branch';
 export { Plural } from './components/branches/Plural';
 export { Derive } from './components/derivation/Derive';
-export { GtInternalTranslateJsx, RscT, T } from './components/translation/T';
+export { GtInternalTranslateJsx, T } from './components/translation/T';
+export { RscT } from './components/translation/T.rsc';
 export { Currency } from './components/variables/Currency';
 export { DateTime } from './components/variables/DateTime';
 export { Num } from './components/variables/Num';
@@ -20,6 +21,9 @@ export { useGT } from './hooks/useGT';
 export { useMessages } from './hooks/useMessages';
 export { useTranslations } from './hooks/useTranslations';
 export { getFormatLocales, useFormatLocales } from './hooks/utils';
+export { createRenderPipeline } from './utils/rendering/createRenderPipeline';
+export type { RenderPipeline } from './utils/rendering/createRenderPipeline';
+export type { RenderPreparedT } from './utils/translation/prepareT.shared';
 
 // ===== Functions ===== //
 export { getTranslationsSnapshot } from './functions/helpers/getTranslationsSnapshot';
