+++ packages/react-core/src/hooks/utils.ts
@@ -1,6 +1,5 @@
 import { useMemo } from 'react';
 import { useEnableI18n, useLocale } from './condition-store';
-import { getReadonlyConditionStoreWithFallback } from '../condition-store/singleton-operations';
 import { getFormatLocales } from './utils/getFormatLocales';
 import { getI18nConfig } from 'gt-i18n/internal';
 
@@ -28,16 +27,4 @@ export function useShouldTranslate(): boolean {
   const enableI18n = useEnableI18n();
   const locale = useLocale();
   return enableI18n && getI18nConfig().requiresTranslation(locale);
-}
-
-/**
- * Returns true if (1) i18n enabled and (2) translation is required
- */
-export function getShouldTranslate(): boolean {
-  const conditionStore = getReadonlyConditionStoreWithFallback();
-  const i18nConfig = getI18nConfig();
-
-  const enableI18n = conditionStore.getEnableI18n();
-  const locale = conditionStore.getLocale();
-  return enableI18n && i18nConfig.requiresTranslation(locale);
 }
