+++ packages/react/src/__tests__/context-rsc.test.ts
@@ -9,18 +9,13 @@ describe('gt-react/context react-server surface', () => {
     expect(mod.DateTime).toBeTypeOf('function');
     expect(mod.Derive).toBeTypeOf('function');
     expect(mod.GtInternalDerive).toBeTypeOf('function');
-    expect(mod.GtInternalCurrency).toBeTypeOf('function');
-    expect(mod.GtInternalDateTime).toBeTypeOf('function');
-    expect(mod.GtInternalNum).toBeTypeOf('function');
-    expect(mod.GtInternalPlural).toBeTypeOf('function');
-    expect(mod.GtInternalRelativeTime).toBeTypeOf('function');
-    expect(mod.GtInternalVar).toBeTypeOf('function');
     expect(mod.Num).toBeTypeOf('function');
     expect(mod.Plural).toBeTypeOf('function');
     expect(mod.RelativeTime).toBeTypeOf('function');
     expect(mod.RscT).toBeUndefined();
     expect(mod.T).toBeTypeOf('function');
     expect(mod.Var).toBeTypeOf('function');
+    expect(mod.GtInternalVar).toBeTypeOf('function');
     expect(mod.getFormatLocales).toBeTypeOf('function');
     expect(mod.getPluralBranch).toBeTypeOf('function');
     expect(mod.GTProvider).toBeTypeOf('function');
