+++ packages/react/src/context.client.ts
@@ -48,7 +48,13 @@ export {
   getTranslationsSnapshot,
   getReactI18nCache,
   setReactI18nCache,
+  createRenderPipeline,
   t,
   // ===== Setup ===== //
   internalInitializeGTSRA as initializeGT,
+} from '@generaltranslation/react-core/context';
+
+export type {
+  RenderPipeline,
+  RenderPreparedT,
 } from '@generaltranslation/react-core/context';
